Add WEEKS support and split budget_tracker into a package
Convert budget_tracker.py into a package with two modules:
- __init__.py: BudgetTracker base class, BudgetWindow, window
  computation helpers, singleton factory
- in_memory.py: InMemoryBudgetTracker implementation

Add WEEKS duration support in _compute_window_start and
_compute_window_end, using epoch-aligned 7-day intervals.
